+++ Demo/src/main/java/com/braintreepayments/demo/CardFragment.java
@@ -445,7 +445,7 @@ public class CardFragment extends BaseFragment implements OnCardFormSubmitListen
         threeDSecureRequest.setV2UiCustomization(v2UiCustomization);
         threeDSecureRequest.setV1UiCustomization(v1UiCustomization);
 
-        threeDSecureRequest.setUIType(ThreeDSecureRequest.BOTH);
+        threeDSecureRequest.setUiType(ThreeDSecureRequest.BOTH);
 
         ArrayList renderTypes = new ArrayList<>();
         renderTypes.add(ThreeDSecureRequest.OTP);